src/test/java/com/synechron/wordcount/api/WordCounterControllerTest.java
package com.synechron.wordcount.api;

import com.fasterxml.jackson.core.JsonProcessingException;
import com.fasterxml.jackson.databind.ObjectMapper;
import com.fasterxml.jackson.databind.ObjectWriter;
import com.synechron.Word;
import com.synechron.wordcount.service.WordCounterService;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.boot.test.json.JacksonTester;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;

import java.util.Locale;

import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;

@ExtendWith(MockitoExtension.class)
class WordCounterControllerTest {

    private MockMvc mvc;

    @Mock
    private WordCounterService wordCounterService;

    @InjectMocks
    private WordCounterController wordCounterController;

    @BeforeEach
    public void setup() {
        JacksonTester.initFields(this, new ObjectMapper());
        mvc = MockMvcBuilders.standaloneSetup(wordCounterController).build();
    }

    @Test
    void shouldReturnCountOnGet() throws Exception {

        when(wordCounterService.count("flower")).thenReturn(3);

        mvc.perform(get("/v1/words/flower/count")
                        .accept(MediaType.APPLICATION_JSON))
                .andExpect(MockMvcResultMatchers.status().isOk())
                .andExpect(MockMvcResultMatchers.content().string("{\"count\":3}"));
    }

    @Test
    void shouldReturnCreatedOnPost() throws Exception {

        mvc.perform(post("/v1/words").content(getJsonString(new Word(Locale.ENGLISH, "flower")))
                        .contentType(MediaType.APPLICATION_JSON))
                .andExpect(MockMvcResultMatchers.status().isCreated());
    }

    private String getJsonString(Word word) throws JsonProcessingException {
        ObjectWriter writer = new ObjectMapper().writer().withDefaultPrettyPrinter();
        return writer.writeValueAsString(word);
    }
}